requirements.txt
aiokafka==0.10.0
asyncpg==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
uvloop==0.19.0; platform_system != 'Windows'
urllib3==1.26.18


